requirements.txt
flask==3.1.0
pymongo==4.10.1
python-dotenv==1.0.1
gunicorn==23.0.0
